BackendServer/requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-filter==0.13.0
djangorestframework==3.3.3
gunicorn==19.4.5
Markdown==2.6.6
MySQL-python==1.2.5
pbr==1.9.1
PyMySQL==0.7.2
pytz==2016.4
requests==2.10.0
six==1.10.0
stevedore==1.12.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
web.py==0.37
whitenoise==3.0
psycopg2==2.6.1

